Ethnicity, age, marital status of residents in Newark according to 1880 US Census.

A dataset containing various attributes of residents in Newark. A historical dataset developed and maintained by Spatial Structures in the Social Sciences, Brown University (contact: Prof. John Logan; john_logan@brown.edu)

data(Newark)

Format

A data frame with 21520 rows and 10 variables:

ID

Unique identifier for observation

X

Geocode: Longitude of the observation

Y

Geocode: Latitude of the observation

X_UTM2

Geocode: X coordinate of the observation in Universal Traverse Mercator, false origin, and jiggled to create unique coordinates for observations

Y_UTM2

Geocode: Y coordinate of the observation in Universal Traverse Mercator, false origin, and jiggled to create unique coordinates for observations

NW

Ethnicity indicator: American

IRISH

Ethnicity indicator: Irish

GERMAN

Ethnicity indicator: German

under30

Age of respondent is under 30

mar

Indicator variable for marital status = married

usborn

Respondent was born in the USA

...

Source

https://s4.ad.brown.edu/Projects/UTP/index.htm