A function to compute jointcount test for binomial and multinomial (asymptotic and permutation distributions)

A function to compute Q test for spatial qualitative data

jc.test(formula = NULL,
               data = NULL,
               fx = NULL,
               listw = NULL,
               na.action,
               zero.policy = NULL,
               distr = "asymptotic",
               alternative = "greater",
               control =list())

Arguments

formula
a symbolic description of the factor(s).

data
an (optional) data frame or a sf object with points/multipolygons geometry containing the variable(s) to be tested.

fx
a factor or a matrix of factors in columns

listw
A listw object created for example by nb2listw from spatialreg package; if nb2listw not given, the spatial weights are built using the object given in listw argument (usually an sf object). Default = NULL.

na.action
A function (default options("na.action")), can also be na.omit or na.exclude with consequences for residuals and fitted values. It may be necessary to set zero.policy to TRUE because this subsetting may create no-neighbour observations.

zero.policy
Similar to the corresponding parameter of lagsarlm function in spatialreg package. If TRUE assign zero to the lagged value of zones without neighbours. Default = NULL.

distr
character. Distribution type "asymptotic" (default) or "mc".

control
Optional argument. See Control Argument section.

Value

An object of the class htest

Details

Aquí Antonio escribe una linda historia ....

Control arguments

nsim

number of permutations for get the Monte Carlo distribution. Default = 999

seedinit

seed to select the initial element to star the algorithm to get compute the m-surroundings or to start the simulations.

References

Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, pp. 19-20.

Upton, G., Fingleton, B. 1985 Spatial data analysis by example: point pattern and qualitative data, Wiley, pp. 158–170.

See also

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paezha@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es

Examples

 ## Case 1
 ## Multinomial + Binomial using a sf multipolygon
rm(list = ls())
data("Spain")
f1 <- ~ Older65 + MenWoman
jc1 <- jc.test(formula = f1,
               data = spain.sf,
               distr = "mc",
               alternative = "greater",
               zero.policy = TRUE)
summary(jc1)

#> <div id="bhtlseklmp" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #bhtlseklmp .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #bhtlseklmp .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #bhtlseklmp .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #bhtlseklmp .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #bhtlseklmp .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #bhtlseklmp .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #bhtlseklmp .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #bhtlseklmp .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #bhtlseklmp .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #bhtlseklmp .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #bhtlseklmp .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #bhtlseklmp .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #bhtlseklmp .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #bhtlseklmp .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #bhtlseklmp .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #bhtlseklmp .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #bhtlseklmp .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #bhtlseklmp .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #bhtlseklmp .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #bhtlseklmp .gt_left {
#>   text-align: left;
#> }
#> 
#> #bhtlseklmp .gt_center {
#>   text-align: center;
#> }
#> 
#> #bhtlseklmp .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #bhtlseklmp .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #bhtlseklmp .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #bhtlseklmp .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #bhtlseklmp .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #bhtlseklmp .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 65%;
#> }
#> </style>
#>   <table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="5" class="gt_heading gt_title gt_font_normal gt_bottom_border" style>JoinCount Spatial Tests (Monte Carlo)</th>
#>     </tr>
#>     
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">pairs</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">pvalue</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Joincount</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Expected</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Variance</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <td colspan="5" class="gt_group_heading">Older65 - multinomial - alternative: greater - Monte-Carlo simulation of
#>         join-count statistic (nonfree sampling)</td>
#>     </tr>
#>     <tr><td class="gt_row gt_left">high:high</td>
#> <td class="gt_row gt_right">0.39200</td>
#> <td class="gt_row gt_right">13</td>
#> <td class="gt_row gt_right">12.32</td>
#> <td class="gt_row gt_right">9.50</td></tr>
#>     <tr><td class="gt_row gt_left">low:low</td>
#> <td class="gt_row gt_right">0.52100</td>
#> <td class="gt_row gt_right">9</td>
#> <td class="gt_row gt_right">9.40</td>
#> <td class="gt_row gt_right">8.32</td></tr>
#>     <tr><td class="gt_row gt_left">middle:middle</td>
#> <td class="gt_row gt_right">0.75700</td>
#> <td class="gt_row gt_right">12</td>
#> <td class="gt_row gt_right">13.80</td>
#> <td class="gt_row gt_right">10.17</td></tr>
#>     <tr><td class="gt_row gt_left">low:high</td>
#> <td class="gt_row gt_right">0.23600</td>
#> <td class="gt_row gt_right">26</td>
#> <td class="gt_row gt_right">22.96</td>
#> <td class="gt_row gt_right">16.66</td></tr>
#>     <tr><td class="gt_row gt_left">middle:high</td>
#> <td class="gt_row gt_right">0.70700</td>
#> <td class="gt_row gt_right">26</td>
#> <td class="gt_row gt_right">27.87</td>
#> <td class="gt_row gt_right">18.27</td></tr>
#>     <tr><td class="gt_row gt_left">middle:low</td>
#> <td class="gt_row gt_right">0.43700</td>
#> <td class="gt_row gt_right">25</td>
#> <td class="gt_row gt_right">24.65</td>
#> <td class="gt_row gt_right">17.15</td></tr>
#>     <tr><td class="gt_row gt_left">Jtot</td>
#> <td class="gt_row gt_right">0.41900</td>
#> <td class="gt_row gt_right">77</td>
#> <td class="gt_row gt_right">75.48</td>
#> <td class="gt_row gt_right">21.47</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="5" class="gt_group_heading">MenWoman - binomial - alternative: greater - Monte-Carlo simulation of join-count statistic</td>
#>     </tr>
#>     <tr><td class="gt_row gt_left">men-men</td>
#> <td class="gt_row gt_right">0.52000</td>
#> <td class="gt_row gt_right">60</td>
#> <td class="gt_row gt_right">60.42</td>
#> <td class="gt_row gt_right">24.75</td></tr>
#>     <tr><td class="gt_row gt_left">woman-woman</td>
#> <td class="gt_row gt_right">0.01700</td>
#> <td class="gt_row gt_right">13</td>
#> <td class="gt_row gt_right">7.06</td>
#> <td class="gt_row gt_right">5.82</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

f2 <- ~ MenWoman + Coast
jc2 <- jc.test(formula = f2,
               data = spain.sf,
               distr = "mc",
               zero.policy = TRUE)
summary(jc2)

#> <div id="hbinosngnf" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #hbinosngnf .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #hbinosngnf .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #hbinosngnf .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #hbinosngnf .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #hbinosngnf .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #hbinosngnf .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #hbinosngnf .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #hbinosngnf .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #hbinosngnf .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #hbinosngnf .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #hbinosngnf .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #hbinosngnf .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #hbinosngnf .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #hbinosngnf .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hbinosngnf .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #hbinosngnf .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #hbinosngnf .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #hbinosngnf .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #hbinosngnf .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #hbinosngnf .gt_left {
#>   text-align: left;
#> }
#> 
#> #hbinosngnf .gt_center {
#>   text-align: center;
#> }
#> 
#> #hbinosngnf .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #hbinosngnf .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #hbinosngnf .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #hbinosngnf .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #hbinosngnf .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #hbinosngnf .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 65%;
#> }
#> </style>
#>   <table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="5" class="gt_heading gt_title gt_font_normal gt_bottom_border" style>JoinCount Spatial Tests (Monte Carlo)</th>
#>     </tr>
#>     
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">pairs</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">pvalue</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Joincount</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Expected</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Variance</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <td colspan="5" class="gt_group_heading">MenWoman - binomial - alternative: greater - Monte-Carlo simulation of join-count statistic</td>
#>     </tr>
#>     <tr><td class="gt_row gt_left">men-men</td>
#> <td class="gt_row gt_right">0.54300</td>
#> <td class="gt_row gt_right">60</td>
#> <td class="gt_row gt_right">60.61</td>
#> <td class="gt_row gt_right">24.43</td></tr>
#>     <tr><td class="gt_row gt_left">woman-woman</td>
#> <td class="gt_row gt_right">0.01750</td>
#> <td class="gt_row gt_right">13</td>
#> <td class="gt_row gt_right">7.07</td>
#> <td class="gt_row gt_right">6.02</td></tr>
#>     <tr class="gt_group_heading_row">
#>       <td colspan="5" class="gt_group_heading">Coast - binomial - alternative: greater - Monte-Carlo simulation of join-count statistic</td>
#>     </tr>
#>     <tr><td class="gt_row gt_left">no-no</td>
#> <td class="gt_row gt_right">0.00100</td>
#> <td class="gt_row gt_right">64</td>
#> <td class="gt_row gt_right">36.70</td>
#> <td class="gt_row gt_right">19.40</td></tr>
#>     <tr><td class="gt_row gt_left">yes-yes</td>
#> <td class="gt_row gt_right">0.91800</td>
#> <td class="gt_row gt_right">14</td>
#> <td class="gt_row gt_right">18.94</td>
#> <td class="gt_row gt_right">13.16</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# Case 2:
## Multinomial using a sf multipoint
rm(list = ls())
data("FastFood")
f1 <- ~ Type
jc3 <- jc.test(formula = f1,
               data = FastFood.sf,
               distr = "asymptotic",
               control = list(knn = 6))

#> Warning: bounding box has potentially an invalid value range for longlat data

#> Warning: bounding box has potentially an invalid value range for longlat data

#> Warning: knearneigh: coordinates are not geographical: longlat argument wrong

 summary(jc3)

#> <div id="xflnarjvfs" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #xflnarjvfs .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #xflnarjvfs .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 6px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #xflnarjvfs .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #xflnarjvfs .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #xflnarjvfs .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #xflnarjvfs .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #xflnarjvfs .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #xflnarjvfs .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #xflnarjvfs .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #xflnarjvfs .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #xflnarjvfs .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #xflnarjvfs .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #xflnarjvfs .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #xflnarjvfs .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #xflnarjvfs .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #xflnarjvfs .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #xflnarjvfs .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #xflnarjvfs .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #xflnarjvfs .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #xflnarjvfs .gt_left {
#>   text-align: left;
#> }
#> 
#> #xflnarjvfs .gt_center {
#>   text-align: center;
#> }
#> 
#> #xflnarjvfs .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #xflnarjvfs .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #xflnarjvfs .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #xflnarjvfs .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #xflnarjvfs .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #xflnarjvfs .gt_footnote_marks {
#>   font-style: italic;
#>   font-weight: normal;
#>   font-size: 65%;
#> }
#> </style>
#>   <table class="gt_table">
#>   <thead class="gt_header">
#>     <tr>
#>       <th colspan="6" class="gt_heading gt_title gt_font_normal gt_bottom_border" style>JoinCount Spatial Tests (asymptotic)</th>
#>     </tr>
#>     
#>   </thead>
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1">pairs</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">z-value</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">pvalue</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Joincount</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Expected</th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_right" rowspan="1" colspan="1">Variance</th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr class="gt_group_heading_row">
#>       <td colspan="6" class="gt_group_heading">Type - multinomial - alternative: greater - Join count test under nonfree sampling</td>
#>     </tr>
#>     <tr><td class="gt_row gt_left">H:H</td>
#> <td class="gt_row gt_right">&minus;0.23</td>
#> <td class="gt_row gt_right">0.58942</td>
#> <td class="gt_row gt_right">255.5</td>
#> <td class="gt_row gt_right">258.05</td>
#> <td class="gt_row gt_right">127.06</td></tr>
#>     <tr><td class="gt_row gt_left">P:P</td>
#> <td class="gt_row gt_right">0.30</td>
#> <td class="gt_row gt_right">0.38114</td>
#> <td class="gt_row gt_right">317.0</td>
#> <td class="gt_row gt_right">313.38</td>
#> <td class="gt_row gt_right">143.47</td></tr>
#>     <tr><td class="gt_row gt_left">S:S</td>
#> <td class="gt_row gt_right">&minus;1.82</td>
#> <td class="gt_row gt_right">0.96574</td>
#> <td class="gt_row gt_right">283.5</td>
#> <td class="gt_row gt_right">305.14</td>
#> <td class="gt_row gt_right">141.19</td></tr>
#>     <tr><td class="gt_row gt_left">P:H</td>
#> <td class="gt_row gt_right">0.02</td>
#> <td class="gt_row gt_right">0.49358</td>
#> <td class="gt_row gt_right">571.0</td>
#> <td class="gt_row gt_right">570.72</td>
#> <td class="gt_row gt_right">304.67</td></tr>
#>     <tr><td class="gt_row gt_left">S:H</td>
#> <td class="gt_row gt_right">0.05</td>
#> <td class="gt_row gt_right">0.48128</td>
#> <td class="gt_row gt_right">564.0</td>
#> <td class="gt_row gt_right">563.18</td>
#> <td class="gt_row gt_right">301.61</td></tr>
#>     <tr><td class="gt_row gt_left">S:P</td>
#> <td class="gt_row gt_right">1.08</td>
#> <td class="gt_row gt_right">0.14022</td>
#> <td class="gt_row gt_right">640.0</td>
#> <td class="gt_row gt_right">620.53</td>
#> <td class="gt_row gt_right">325.50</td></tr>
#>     <tr><td class="gt_row gt_left">Jtot</td>
#> <td class="gt_row gt_right">0.90</td>
#> <td class="gt_row gt_right">0.18350</td>
#> <td class="gt_row gt_right">1775.0</td>
#> <td class="gt_row gt_right">1,754.43</td>
#> <td class="gt_row gt_right">519.87</td></tr>
#>   </tbody>
#>   
#>   
#> </table>
#> </div>

# Examples function joincount.test
data(oldcol)
HICRIME <- cut(COL.OLD$CRIME, breaks=c(0,35,80), labels=c("low","high"))
names(HICRIME) <- rownames(COL.OLD)
jc4 <- jc.test(fx = HICRIME,
               listw = nb2listw(COL.nb,
               style="B"))

#> Error in jc.test(fx = HICRIME, listw = nb2listw(COL.nb, style = "B")): object 'data.name' not found

 jc5 <- jc.test(fx = HICRIME,
                listw = nb2listw(COL.nb, style="B"),
                distr = "'mc")

#> Error in jc.test(fx = HICRIME, listw = nb2listw(COL.nb, style = "B"),     distr = "'mc"): object 'data.name' not found

 HICRIME <- cut(COL.OLD$CRIME, breaks=c(0,35,80), labels=c("low","high"))
 names(HICRIME) <- rownames(COL.OLD)
 jc6 <- jc.test(fx = HICRIME,
                listw = nb2listw(COL.nb,
                                 style="B"))

#> Error in jc.test(fx = HICRIME, listw = nb2listw(COL.nb, style = "B")): object 'data.name' not found