A function to generate m-surroundings

This function obtains the m-surroundings by selecting the m-1 nearest neighbors of each observation, allowing for a degree of overlap of r.

m.surround(x, m, r = 1, distance = "Euclidean", control = list())

Arguments

x
input sf object with points/multipolygons geometry or matrix of spatial coordinates

m
dimension of m-surrounding

r
maximum overlapping between any two m-surroundings.

distance
character. For Cartesian coordinates only: one of Euclidean, Hausdorff or Frechet; for geodetic coordinates, Great Circle distances are computed.

control
optional. Parameters for pruning m-surroundings with the aim of remove m-surroundings whose elements are too far.

Value

A list of class list and m_surr containing the following components:

ms	a matrix. Each row is a m-surrounding.

R	total number of observations.

mdtms	xxxxxxxxxxxxxxxxxxx.

mscbl	xxxxxxxxxxxxxxxxxxx.

rowexcl	xxxxxxxxxxxxxxxxxxx.

N	total number of symbolized observations.

m	length of the m-surroundings.

r	overlapping degree.

initobs	element to star the generation of m-surroundings. default initobs=1.

distance	distance select to prune the m-surroundings.

m	length of the m-surroundings.

x	the input "x" as sf-object.

Details

Aquí Antonio escribe una linda historia ...

Control arguments

distance

character to select the type of distance. Default = "Euclidean" for Cartesian coordinates only: one of Euclidean, Hausdorff or Frechet; for geodetic coordinates, great circle distances are computed (see sf::st_distance())

dtmaxabs

xxxxxxxxxxxxxxxxxxxxxxxxxxxx

dtmaxpc

xxxxxxxxxxxxxxxxxxxxxxxxxxxx

dtmaxknn

Elimina m-surrounding donde alguno de los elementos no está entre los knn más próximos

References

Ruiz M, López FA, A Páez. (2010). Testing for spatial association of qualitative data using symbolic dynamics. Journal of Geographical Systems. 12 (3) 281-309

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paezha@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es

Examples