Method for class m_surr

A function to plots the m-surround give an object of the class m_surr obtain with the code m.surround.
The plot() function allows the user view the configuration of the m-surrounding. The argument type select the type o visualization.
The print() print the matrix of the m-surrounding.
. The summary give information about the characteristics of th m-surrounding
.

# S3 method for m_surr
summary(object, ...)

# S3 method for m_surr
plot(x, type = 1, ...)

# S3 method for m_surr
print(x, ...)

Arguments

object
object of class m_surr. 2 plot W matrix with network

...
further arguments passed to or from other methods.

x
object of class m_surr

type
numeric. 1 (default) to get the plot with igraph.

References

Ruiz, M., López, F., and Páez, A. (2010). A test for global and local homogeneity of categorical data based on spatial runs. Geographical Analysis.

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paezha@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es

Examples

# Obtain m-surroundings with degree of overlapping r
rm(list = ls())
N <- 100
cx <- runif(N)
cy <- runif(N)
x <- cbind(cx,cy)
m = 4
r = 2
msurr_points <- m.surround(x = x, m = m, r = r,control = list(dtmaxabs = 0.5))

#> 
#>  Threshold distance:  0.5
#>  Number of m-surroundings excluded for exceeding
#>         the threshold distance:  1 
#> 
#>  Index of spatial observations excluded:  18 

plot(msurr_points, type = 1)

#> Error in plot.m_surr(msurr_points, type = 1): object 'm' not found

plot(msurr_points, type = 2)

#> Error in plot.m_surr(msurr_points, type = 2): object 'm' not found

print(msurr_points)

#>       [,1] [,2] [,3] [,4]
#>  [1,]    1   18   75   15
#>  [2,]    4   51   69   28
#>  [3,]    8   10   21   99
#>  [4,]    9   37   84   79
#>  [5,]   12   41   61  100
#>  [6,]   14   77    9   79
#>  [7,]   16   56   72   93
#>  [8,]   19   94   16   56
#>  [9,]   21   99   65   89
#> [10,]   28   53   71   40
#> [11,]   31   59   86   55
#> [12,]   33   67   54   78
#> [13,]   34   42   63   13
#> [14,]   35   32   33   78
#> [15,]   36   11   46    3
#> [16,]   38   13   88   58
#> [17,]   39   20   12   30
#> [18,]   43   22   52    7
#> [19,]   45   17   35   32
#> [20,]   46    3   43    7
#> [21,]   48   98   49    3
#> [22,]   49   25    4   51
#> [23,]   50   44   97   85
#> [24,]   52    7   34   13
#> [25,]   54   74   78   27
#> [26,]   57   64   91   23
#> [27,]   60   30   50   97
#> [28,]   61    5  100   66
#> [29,]   62   26   48   58
#> [30,]   63    2   19   16
#> [31,]   65    6   89   48
#> [32,]   68   87   57   91
#> [33,]   69   24   36   80
#> [34,]   70   95   68   62
#> [35,]   71   82   40   85
#> [36,]   72   47   38   93
#> [37,]   75   15   31   82
#> [38,]   76   80   39   20
#> [39,]   78   90   83   27
#> [40,]   79   73   92   93
#> [41,]   83   27    8   62
#> [42,]   84   81   76   92
#> [43,]   86   55   45   17
#> [44,]   88   58   14   77
#> [45,]   89   96   70   57
#> [46,]   91   23   62   26
#> [47,]   97   29   28   53
#> [48,]  100   66   60   85

rm(list = ls())
data("FastFood")
m = 6
r = 1
msurr_points <-  m.surround(x = FastFood.sf, m = m, r = r, distance = "Euclidean", control = list(dtmaxpc = .01))

#> 
#>  Threshold distance:  3.596891
#>  Number of m-surroundings excluded for exceeding
#>         the threshold distance:  14 
#> 
#>  Index of spatial observations excluded:  72 113 27 32 80 134 136 160 23 38 55 73 104 107 

plot(msurr_points, type = 1)

#> Error in plot.m_surr(msurr_points, type = 1): object 'm' not found

plot(msurr_points, type = 2)

#> Error in plot.m_surr(msurr_points, type = 2): object 'm' not found

print(msurr_points)

#>        [,1] [,2] [,3] [,4] [,5] [,6]
#>   [1,]    1  809  868  300  473  620
#>   [2,]    9  568  422  875  877   10
#>   [3,]   10  763  739  706  487  857
#>   [4,]   11  583  379  421  316  710
#>   [5,]   16  843  296  519   28   11
#>   [6,]   22  572  129  874   20  182
#>   [7,]   30  299  290  458  786  812
#>   [8,]   33  438  513  829  322   90
#>   [9,]   34  566  321  577  151  250
#>  [10,]   35  818  396  435  267  767
#>  [11,]   48  854  867  431  750  714
#>  [12,]   49  212  450  536  705  308
#>  [13,]   55  492  424   59  148  824
#>  [14,]   56  153  463  482  715  111
#>  [15,]   64  426  219  285  171  351
#>  [16,]   66  252  656  762  820  278
#>  [17,]   79  257  246   77  733  328
#>  [18,]   90  484  796  362  556  689
#>  [19,]   96  233  856  794  779  859
#>  [20,]  100  836   88  158  412  335
#>  [21,]  101   78  834  802  823  311
#>  [22,]  106  243   21  242  103  678
#>  [23,]  115  761  808  480  797  776
#>  [24,]  116  276  339  512  591  525
#>  [25,]  130  374  625   76  416   64
#>  [26,]  135  347  841  145  663  217
#>  [27,]  139  567  195  159  866  284
#>  [28,]  143  774  589  349  662    9
#>  [29,]  147  382  253  830  683  197
#>  [30,]  154   37  141  224  211  467
#>  [31,]  157  232  249  137  600   34
#>  [32,]  166  734  240  605   87  291
#>  [33,]  172  585   45  768  238  221
#>  [34,]  173  575  404  403  592  318
#>  [35,]  183   31  713   54  855   96
#>  [36,]  197  293  209  815  385  268
#>  [37,]  202  260  320  310  510  231
#>  [38,]  203   23  717  534  481  405
#>  [39,]  207  870  126  441  468  150
#>  [40,]  216   53  186  372   85  738
#>  [41,]  217  333  292  828  787  735
#>  [42,]  218   19  375  811  778  876
#>  [43,]  220  326  684  569  674  447
#>  [44,]  221  775  730  119  746  323
#>  [45,]  230  527  236   46  189  218
#>  [46,]  231  389  553  123  680  419
#>  [47,]  250  483  533  764  668  690
#>  [48,]  259  765  271  418  822  799
#>  [49,]  261   69  417  114  850  202
#>  [50,]  262  649  241  621  237  531
#>  [51,]  268  331  423  634  210  707
#>  [52,]  275  256  355  840  393  517
#>  [53,]  278  557  754  370  169  388
#>  [54,]  283  214  632  518  509   49
#>  [55,]  284  227   42  702   80  831
#>  [56,]  287  485  191  504  785  547
#>  [57,]  291   41  234  136  524  853
#>  [58,]  295  807  165  716  576  216
#>  [59,]  298  672  528  427  697  366
#>  [60,]  302  146  782  792  693  173
#>  [61,]  303  654  486  562  772  832
#>  [62,]  304  133  478  645  619  570
#>  [63,]  307  542  790   14  270  337
#>  [64,]  308  545   38  615  398  157
#>  [65,]  311  623  729  651  626  595
#>  [66,]  315  161  312  345  144  629
#>  [67,]  327  269  391  201  741   35
#>  [68,]  328   75  365  109  163  756
#>  [69,]  329  444  281  610   44  443
#>  [70,]  332  340  177  633  454  303
#>  [71,]  335  400  709   93  554  101
#>  [72,]  337  872    3  755  688  230
#>  [73,]  341  675  167  239  105  134
#>  [74,]  350  642  222   92  104   79
#>  [75,]  351  297  636   40  744   30
#>  [76,]  366  670  607   47   57  135
#>  [77,]  368  470  564  846  507  810
#>  [78,]  376  616  793  635  760  743
#>  [79,]  388   65  520  174   94  530
#>  [80,]  394  280  436  745  131  302
#>  [81,]  395  277  440   61  254  851
#>  [82,]  397  428  113   32  858  220
#>  [83,]  405  587  677  413  737  295
#>  [84,]  419   67  317   18  245  821
#>  [85,]  425  378  686  453  725  283
#>  [86,]  434  606  731   81  781  474
#>  [87,]  437   95  721    5  805  537
#>  [88,]  443  179  386  198  193  307
#>  [89,]  446  247  806  671  665  304
#>  [90,]  447  415  384  588  593  658
#>  [91,]  457  228   29  336  817  394
#>  [92,]  461  758  455  631  363  718
#>  [93,]  465  701  204  766  314   33
#>  [94,]  467  258  641  622  199  259
#>  [95,]  471  451  442  226  306  548
#>  [96,]  477  703  188  653  353  275
#>  [97,]  489  652   86  354  609  376
#>  [98,]  515  357  771  727  602  207
#>  [99,]  517  835  704  469    4  627
#> [100,]  522  140  390  411   17  172
#> [101,]  525  865  586  647   24   22
#> [102,]  530   15  279  456  511  263
#> [103,]  532  272  827  168  546  751
#> [104,]  537   51  402  614  804  397
#> [105,]  541  392  475  728   36  139
#> [106,]  547  604  813  380  176  659
#> [107,]  548  560  508  581   13  861
#> [108,]  561   25  596  845  420  740
#> [109,]  570  466  657  488  346  147
#> [110,]  578  816  192  732  215  425
#> [111,]  582  844  640  719  121  603
#> [112,]  595  255  187  495  574  532
#> [113,]  603  445  497  535   70  100
#> [114,]  613  289  406  160  661  679
#> [115,]  620  348  274  526  848  261
#> [116,]  627   74  780  342  313  315
#> [117,]  629  493  597  676  847  287
#> [118,]  630  502   73  747  499  130
#> [119,]  644  814    8  699  660  341
#> [120,]  655  559  784   50  213  724
#> [121,]  658  871  523   97  381  203
#> [122,]  659  696  543  162  549  368
#> [123,]  678  637  358   27  496  515
#> [124,]  681  112  452  156  692  183
#> [125,]  689  361  694  309   58  613
#> [126,]  690  205  685  770   99  166
#> [127,]  695  646  618  860   26  338
#> [128,]  707  432  149  753    2  630
#> [129,]  708  800  122  356  825  298
#> [130,]  710  837  759  190  503  578
#> [131,]  711  265  552  712  849  561
#> [132,]  714  666  538  178  650  769
#> [133,]  718  638  414  325  514  489
#> [134,]  723  410  244   84  752   66
#> [135,]  724  459  142  359   63  819
#> [136,]  735  624  448  748  544  695
#> [137,]  738   89  612  264  208  327
#> [138,]  740  164  498  377  185  490
#> [139,]  743  842  691  430  590  106
#> [140,]  751  869  235  369  682  437
#> [141,]  756  125  343  383   43  446
#> [142,]  767  305  862  555   68  711
#> [143,]  769  791  433  127  826  465
#> [144,]  776  439  673  501  371  477
#> [145,]  795  225  138  460  720   48
#> [146,]  799    7  617  206  196  833
#> [147,]  812  273  667  248  722  262
#> [148,]  819  521  324  229  118  471
#> [149,]  821  407  563  409  742  522
#> [150,]  824  565  301  373  223  873
#> [151,]  831  282  155  124  573  350
#> [152,]  832  251  472   91  462  387
#> [153,]  833  540  580  117  788  839
#> [154,]  839  194  175  852  736  395
#> [155,]  851  664  698   82  344  115
#> [156,]  853  401  749  367  200  143
#> [157,]  857  170  464   39  584  332
#> [158,]  859  539  102  783   71  708
#> [159,]  861  777  360  120  449  681
#> [160,]  873   83   60  516  757  669
#> [161,]  876  132  601  611  505  116