A function to order the elements og the m_i-subrrounds

An auxiliary function. In the case of obtaining the list of neighbors of class nb or poly2nb, it is necessary to reorder the elements based on distance and/or angle.

nb2nb_order(listw = listw, sf = sf)

Arguments

listw
an object of the nb class.

sf
the sf object used to get the listw .

Value

An object of the nb class with the elements in order.

Details

Sort the elements of a list nb. First by distance and

See also

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paez@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es

@references

Ruiz, M., López, F., and Páez, A. (2010). A test for global and local homogeneity of categorical data based on spatial runs. Geographical Analysis.

Examples

# With a sf object (irregular lattice)
library(sf)
fname <- system.file("shape/nc.shp", package="sf")
nc <- st_read(fname)
#> Reading layer `nc' from data source 
#>   `/home/runner/work/_temp/Library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
listw <- spdep::poly2nb(as(nc,"Spatial"), queen = FALSE)
listw.order <- nb2nb_order(listw = listw, sf = nc)

# With a sf object (regular lattice: hexagons)
sfc = st_sfc(st_polygon(list(rbind(c(0,0), c(1,0), c(1,1), c(0,1), c(0,0)))))
hexs <- st_make_grid(sfc, cellsize = 0.1, square = FALSE)
hexs.sf <- st_sf(hexs)
listw  <- poly2nb(as(hexs.sf, "Spatial"), queen = FALSE)
listw.order <- nb2nb_order(listw = listw, sf = hexs.sf)