Compute the similarity test.

This function compute the similarity test for categorical spatial process.

similarity.test(formula = NULL, data = NULL, na.action, fx = NULL, listw = listw,
alternative = "two.sided", distr = "asymptotic", nsim = NULL, control = list())

Arguments

formula
a symbolic description of the factor (optional).

data
an (optional) data frame or a sf object containing the variable to testing for.

fx
a factor (optional).

alternative
a character string specifying the type of cluster, must be one of "High" (default), "Both" or "Low".

distr
A string. Distribution of the test "asymptotic" (default) or "bootstrap"

nsim
Number of permutations.

control
List of additional control arguments.

Value

A object of the htest

data.name	a character string giving the names of the data.

statistic	Value of the similarity test

N	total number of observations.

Zmlc	Elements in the Most Likelihood Cluster.

alternative	a character string describing the alternative hypothesis.

p.value	p-value of the similarity test

similiarity.mc	values of the similarity test in each permutation.

Details

Escribir aquí cosas

Control arguments

seedinit	Numerical value for the seed (only for boot version). Default value seedinit=123

References

Farber, S., Marin, M. R., & Paez, A. (2015). Testing for spatial independence using similarity relations. Geographical Analysis. 47(2), 97-120.

See also

Author

Fernando López	fernando.lopez@upct.es

Román Mínguez	roman.minguez@uclm.es

Antonio Páez	paezha@gmail.com

Manuel Ruiz	manuel.ruiz@upct.es

Examples

rm(list = ls())
N <- 100
cx <- runif(N)
cy <- runif(N)
listw <- knearneigh(cbind(cx,cy), k = 3)
p <- c(1/4,1/4,1/4,1/4)
rho <- 0.5
fx <- dgp.spq(p = p, listw = listw, rho = rho)
W <- (nb2mat(knn2nb(listw)) >0)*1
similarity <- similarity.test(fx = fx, data = FastFood.sf, listw = listw)
print(similarity)
#> 
#> 	Similarity test of spatial dependence for qualitative data.
#> 	Distribution: asymptotic
#> 
#> data:  fx
#> Similarity-test = 1.7228, p-value = 0.08493
#> alternative hypothesis: two.sided
#> 

# Case 2: SRQ test with formula, a sf object (points) and knn
rm(list = ls())
data("FastFood")
coor <- cbind(FastFood.sf$Lon,FastFood.sf$Lat)
listw <- spdep::knearneigh(coor, k = 4)
formula <- ~ Type
similarity <- similarity.test(formula = formula, data = FastFood.sf, listw = listw)
print(similarity)
#> 
#> 	Similarity test of spatial dependence for qualitative data.
#> 	Distribution: asymptotic
#> 
#> data:  Type
#> Similarity-test = -5.4476, p-value = 5.105e-08
#> alternative hypothesis: two.sided
#> 

# Case 3:
rm(list = ls())
data("Spain")
listw <- spdep::poly2nb(as(spain.sf,"Spatial"), queen = FALSE)
formula <- ~ MenWoman
similarity <- similarity.test(formula = formula, data = spain.sf, listw = listw)
#> Warning: Out-of-range p-value: reconsider test arguments
print(similarity)
#> 
#> 	Similarity test of spatial dependence for qualitative data.
#> 	Distribution: asymptotic
#> 
#> data:  MenWoman
#> Similarity-test = NaN, p-value = NA
#> alternative hypothesis: two.sided
#>