spqdata: A package for the exploratory analysis of cross-sectional spatial qualitative data

A package for the exploratory analysis of qualitative data with georeferences.

Details

Functionality included in spqdata are:

Datasets

spqdata includes two different datasets: spain and Fastfood. These sets are used to illustrate the capabilities of different functions. Briefly, their main characteristics are the following

The FastFood An object of class sf with georeferenced points of a selection of fast food restaurants in Toronto, Canada.

The spain.sp An object of class sf with the boundaries of Spanish provinces and several socio-economic variables.

References

Breusch T, Pagan A (1980). The Lagrange multiplier test and its applications to model specification in econometrics. Review of Economic Studies 47: 239-254.

LeSage, J., and Pace, R. K. (2009). Introduction to spatial econometrics. Chapman and Hall/CRC.

López, F.A., Mur, J., and Angulo, A. (2014). Spatial model selection strategies in a SUR framework. The case of regional productivity in EU. Annals of Regional Science, 53(1), 197-220.

López, F.A., Martínez-Ortiz, P.J., and Cegarra-Navarro, J.G. (2017). Spatial spillovers in public expenditure on a municipal level in Spain. Annals of Regional Science, 58(1), 39-65.

Mur, J., López, F., and Herrera, M. (2010). Testing for spatial effects in seemingly unrelated regressions. Spatial Economic Analysis, 5(4), 399-440.

See also

Author

Maintainer: Antonio Páez paezha@mcmaster.ca (ORCID)

Authors:

Román Minguez Salido Roman.Minguez@uclm.es (ORCID)

Fernando López Hernandez fernando.lopez@upct.es (ORCID)

Manuel Ruiz Marin manuel.ruiz@upct.es (ORCID)